Some small changes to selecting which constants to use. In line with the document now. Accounts for day / night.
Pretty complete, unsure if this can be used for anything super useful, but we'll see! Now onto either decay with this model or finding a better model to use. Nonetheless this was a good basic model to build.

Temporarily allowing git to version the build files so node doesn't need to be setup!